requirements.txt
python-dateutil
numpy
pandas
matplotlib